Use gRPC middleware to consistently record client-side metrics for calls to the ingester.
